% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{cfm}
\alias{cfm}
\title{Confusion Matrix}
\usage{
cfm(motif_found, simulated_sequences, complement = F)
}
\arguments{
\item{simulated_sequences}{list; Output of simulate_sequences}

\item{complement}{logical; Did findamotif return the complemented version of the simulated motif
Change this if you get 0 true positives}

\item{found_motif}{list; The output of getmotifs() or findamotif()}
}
\value{
List of confusion matricies, one for overall and one each for each strand
}
\description{
Confusion Matrix
}
\details{
Note that for a positive match, all of the sequence, strand and position of the motif.
A mismatch in any of these will exclude a predicted +ve true +ve. However the total number of true -ve does not
count each position as one negative but each sequence which does not contain an added motif.

Also note that there could be matches to your simulated motif by chance in random sequence, and as the motif
generation is probabilistic, some sequences with the motif added will not be a good match to the consensus motif.
}
