% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{export_FASTA}
\alias{export_FASTA}
\title{Export character vector to file in FASTA format}
\usage{
export_FASTA(sequences, file)
}
\arguments{
\item{sequences}{named character vector; Sequences to be exported}

\item{file}{string; Filename and path for sequences to be saved in}
}
\value{
Nothing, file written to disk.
}
\description{
Export character vector to file in FASTA format
}
\examples{

dna <- c(seqA="ATGCTAG",seqB="ATCGATGTT",seqC="TCGATCGAT")
export_FASTA(dna, "dna.fasta")
}
