% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{extract_matches}
\alias{extract_matches}
\title{Extract Motif Matches}
\usage{
extract_matches(found_motif, motif = 1)
}
\arguments{
\item{found_motif}{list; The output of getmotifs()}

\item{motif}{integer; Which motif to mask if there is more than one in the found_motif list.}
}
\value{
charachter vector of DNA sequences matching the motif
}
\description{
Extract Motif Matches
}
