% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findamotif.R
\name{findamotif}
\alias{findamotif}
\title{Find the single most enriched motif in a set of DNA sequences}
\usage{
findamotif(seqs, len, scores = NULL, nits = 100, ntries = 1,
  n_for_refine = 1000, prior = NULL, updateprior = 1, plen = 0.9,
  seed = NULL)
}
\arguments{
\item{seqs}{a vector of strings giving the DNA sequences in which to find a motif}

\item{len}{length of motif to find (min=4)}

\item{scores}{a set of regional scores giving weights; e.g. ChIP-Seq enrichment values}

\item{nits}{number of iterations used for motif refinement}

\item{ntries}{usually leave at default, number of motifs to be attempted from list of possible starts}

\item{n_for_refine}{the top n_for_refine scoring regions only are used for motif refinement}

\item{prior}{a vector of length 10 probabilities giving the initial probability of a motif being found across different parts of the sequence from start:end. If left unspecified the initial prior is set at uniform and the algorithm tries to learn where motifs are, e.g. if they are centrally enriched.}

\item{updateprior}{a flag - should the algorithm update (learn) the prior on where the motifs occur within the DNA sequences(default is 1)?}

\item{plen}{a parameter setting the geometric prior on how long each motif found should be. plen=0.05 corresponds to a mean length of 20bp and is the default. Setting plen large penalises longer motifs more}

\item{seed}{integer; seed for random number generation, set this for exactly reproducible results.}
}
\value{
List item with the following items:\cr
Details of input data given:
\itemize{
\item seqs: the vector of input sequences used for finding motifs within
\item trimmedseqs: the vector of input sequences used for finding motifs within, after trimming to shorten long input sequences
}

Details of overall fitted model:
\itemize{
\item scoremat: a matrix giving the pwm (log-scale) for the identified motif after iteration
\item scorematdim: the length of the identified motif, and scoremat is of dimension scorematdimx4
\item prior: a vector of length 10 probabilities giving the inferred probability of a motif being found across different parts of the sequence from start to end.
\item alpha: a vector of probabilities giving the inferred probability of the motif being found within each input region
\item bindmat: a version of scoremat accounting for the background sequence composition
\item background is the inferred background model
}

Details of output for given data:
\itemize{
\item regprobs, regprob are in this case identical vectors giving the probability of the motif occurring in each given input sequence
\item bestpos is a vector giving the best match to the motif in each given input sequence
\item whichregs is a vector showing which input sequences had motifs identified in the final round of sampling of the Gibbs sampler
\item whichpos: for motifs identified in regions described in whichreg, the start positions of motifs identified in the final round of sampling of the Gibbs sampler
\item whichmot: not needed in this case
\item whichstrand: for motifs identified in regions described in whichreg, the strand associated with motifs identified in the final round of sampling of the Gibbs sampler, relative to the input sequence
}
}
\description{
Find the single most enriched motif in a set of DNA sequences
}
\details{
This function identifies a single PWM from an iterative Gibbs sampler described in Altemose et al. eLife 2017. Function 2 can refine multiple motifs further, jointly.

The user must input a set of DNA sequences, a score for each sequence (e.g. an enrichment value or any other score), and a length for an initial motif (e.g. 8 bp) used to seed the algorithm.

There are additional optional parameters.

The program outputs a list of results, including information on the inferred PWM (i.e. motif found), as well as a probabilistic output of which regions contain this motif, and posterior distributions of the other parameters
}
