% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{get_PWM}
\alias{get_PWM}
\title{Calculate Position-Weight Matrix}
\usage{
get_PWM(motif, complement = FALSE)
}
\arguments{
\item{motif}{the output of findamotif()}

\item{complment}{logical; if TRUE the reverse complement PWM is returned}
}
\value{
matrix; a 4xN position-weight matrix where N is the length of the motif
}
\description{
Calculate Position-Weight Matrix
}
