% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{load_sequences}
\alias{load_sequences}
\title{Load DNA sequences from a FASTA file}
\usage{
load_sequences(file, sm.rm = T, flank = 150, proportion_N = 0.02)
}
\arguments{
\item{file}{string; Filename and path of FASTA file of sequences}

\item{sm.rm}{logical; should soft masked bases (lower case) be converted to N
or capitalised, default=TRUE}

\item{flank}{integer; how many bases either side of the center should be loaded}

\item{proportion_N}{numeric; fraction of N bases allowed before the
sequence is removed}
}
\value{
charachter vector of DNA sequences
}
\description{
Load DNA sequences from a FASTA file
}
