% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{mask_motif}
\alias{mask_motif}
\title{Mask occurrences of motif}
\usage{
mask_motif(found_motif, motif = 1)
}
\arguments{
\item{found_motif}{list; The output of getmotifs()}

\item{motif}{integer; Which motif to mask if there is more than one in the found_motif list.}
}
\value{
Same as output of getmotifs() but with motifs removed.
}
\description{
Mask occurrences of motif
}
