% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{pcm2pwm}
\alias{pcm2pwm}
\title{Convert PCM to PWM}
\usage{
pcm2pwm(pcm, pseudocount = NULL)
}
\arguments{
\item{pcm}{matrix; n by 4 matrix of PCM / PFM}

\item{pseudocount}{numeric; value of pseudocount to add to every entry of the PWM (to avoid 0 or 1 counts),
or vector of pseudocounts to add to each nt position}
}
\value{
A position weight matrix
}
\description{
Convert PCM to PWM
}
\examples{
pcm <- t(rbind(matrix(c(rep(10,8),rep(100,8)),nrow=2),
               matrix(c(rep(20,7),0,rep(200,7),0),nrow=2)))
pcm2pwm(pcm)

}
