% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{plot_motif_location}
\alias{plot_motif_location}
\title{Plot location of motifs in input sequences}
\usage{
plot_motif_location(found_motif, linepos = NULL, top_n = NULL,
  linesize = 1)
}
\arguments{
\item{found_motif}{list; The output of getmotifs()}

\item{linepos}{integer; position of line in plot to mark e.g. the TSS}

\item{top_n}{integer; how many input sequences to plot}

\item{linesize}{numeric; thickness of lines representing motif footprints}
}
\value{
A ggplot2 object
}
\description{
Plot location of motifs in input sequences
}
