% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{plot_tomtom_match}
\alias{plot_tomtom_match}
\title{Plot Sequence logo denovo motif vs Tomtom match}
\usage{
plot_tomtom_match(query_motif = NULL, tomtom_match = NULL,
  titles = NULL, yaxis = TRUE)
}
\arguments{
\item{query_motif}{list object output from findamotif()}

\item{tomtom_match}{data.frame; result from reading in the text output of Tomtom
columns required: "Target.ID", "Orientation", "Optimal.offset"}

\item{titles}{character vector; titles to use for each logo if not the default}

\item{yaxis}{logical; If FALSE yaxis values and labels are hidden}
}
\value{
A plot, ggplot2 object
}
\description{
Plot Sequence logo denovo motif vs Tomtom match
}
