% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{pwm2text}
\alias{pwm2text}
\title{Convert PWM matrix to text (character string)}
\usage{
pwm2text(pwm, threshold = 0.7)
}
\arguments{
\item{pwm}{matrix; n by 4 matrix of log scale numeric values}

\item{treshold}{numeric; value between 0.5 and 1 to determine when to display capital vs lowercase letter.}
}
\value{
A charachter string
}
\description{
Convert PWM matrix to text (character string)
}
\examples{
pwm2text(position_weight_matrix)

}
