% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{read_meme}
\alias{read_meme}
\title{Read Meme format Position-Weight Matrix to file}
\usage{
read_meme(file)
}
\arguments{
\item{file}{string; filename including path and extension}
}
\value{
list of length 2, a pwm and a name of motif
}
\description{
Read Meme format Position-Weight Matrix to file
}
