% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{submotif}
\alias{submotif}
\title{Check if a motif is present within another motif}
\usage{
submotif(motif, submotif = rbind(c(0, 1, 0, 0), c(0, 0, 1, 0)),
  ispcm = FALSE, probabilistic = TRUE)
}
\arguments{
\item{motif}{matrix; position weight matrix to check in}

\item{submotif}{matrix; motif to search for, format: position weight matrix, n by 4 (A, C, G, T), (output of pcm2pwm)
(default finds CpG dinucleotides)}

\item{ispcm}{logical; if TRUE will convert pcm to pwm (default: FALSE)}

\item{probabilistic}{logical; if FALSE the sum of the convolution is taken,
if TRUE (default), the product of the sum of each row is returned.
TRUE will only work if the submotif is itself a pwm rather than a generic kernal.}
}
\value{
score of submotif being present in motif
}
\description{
Check if a motif is present within another motif
}
